services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qdrant-client==1.9.1
python-dotenv==1.0.1
openai>=1.30.0
httpx==0.27.0
